feat: send garbage before fragments
This is yet another bizzare behavior that shows up, for some reason, the probability of the first or second packet to NOT be sent after a while is close to 100%, so we send garbage data so we away from this.

The probability of packets dropping after the first is not high, but it is there as it is expected from UDP. This is still a really weird behavior and I don't know whose party is at fault here.

The same behavior can be seen in datagram-test, so it might be quic-go?